requirements/base.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1
psycopg2
pillow
djangorestframework==3.12.2
django-filter==2.4.0
django-cors-headers
djangorestframework-jwt
